made a unidied interface & fixed bugs & refactoring
